Revamped index and stylesheet
Changed the color and the fonts in the css stylesheet to the portfolio. The fonts selected were Raleway & Questviral and the colors range from Wintergreen Dream to Mint Cream in a light gradient. The goal was to give the portfolio more of a natural calmer tone while adding a modern touch that isn't too distracting or busy. I wanted the fonts to give a simple and clean look while also getting straight to the point.